requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Database
psycopg2-binary==2.9.7
pgvector==0.2.4

# Document processing
PyPDF2==3.0.1
PyMuPDF==1.23.8

# Embeddings and ML
sentence-transformers==2.2.2
torch>=2.0.0
transformers==4.35.2
numpy==1.24.3

# API clients
anthropic==0.7.7
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: Progress bars
tqdm==4.66.1